move SearchedUser to tailwind
