+++ chat-react2/src/components/Home/Sidebar/NewChat/NewChatDialog.tsx
@@ -6,7 +6,7 @@ import { Input } from "@/components/shadcn-ui/input";
 import { Label } from "@/components/shadcn-ui/label";
 import { Button } from "@/components/shadcn-ui/button";
 import { Users } from "node_modules/connectycube/dist/types/types";
-import SearchedUser from "./SearchedUser/SearchedUser";
+import SearchedUser from "./SearchedUser";
 import CreateGroupChat from "./CreateGroupChat/CreateGroupChat";
 import Participant from "./CreateGroupChat/Participant/Participant";
 import { useNavigate } from "react-router";
