--- chat-react2/src/components/Home/Sidebar/NewChat/SearchedUser/SearchedUser.scss
@@ -1,49 +1,0 @@
-.found__user {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  position: relative;
-
-  & .add-status {
-    background-color: transparent;
-    color: #007bad;
-    border: 1px solid #36b6eb;
-    padding: 10px 20px;
-    border-radius: 10px;
-    transition: 0.2s;
-    cursor: pointer;
-    &.add:hover {
-      background: #31baf3;
-      color: #ffffff;
-    }
-    &.delete {
-      background: #31baf3;
-      color: #ffffff;
-    }
-  }
-  & .user__main-info {
-    display: flex;
-    flex: 0 0 70%;
-    align-items: center;
-  }
-  & .user__name-info {
-    text-align: left;
-
-    & .user__full-name {
-      color: black;
-    }
-    & .user__login {
-      color: black;
-      font-size: 14px;
-    }
-  }
-  &:not(:last-child) {
-    padding-bottom: 10px;
-  }
-
-  & .user__avatar {
-    width: 50px;
-    height: 50px;
-    margin-right: 15px;
-  }
-}
